examples/langchain/pyproject.toml
[project]
name = "workbench-langchain-examples"
version = "0.1.0"
description = "Workbench LangChain integration examples"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
    "langchain-mcp-adapters>=0.1.11",
    "langgraph>=1.0.0",
    "langchain-anthropic>=1.0.0",
]
